orders/src/routes/__test__/new.test.ts
import mongoose from 'mongoose';
import { app } from '../../app';
import request from 'supertest';
import { Ticket } from '../../models/ticket';
import { Order, OrderStatus } from '../../models/order';
import { natsWrapper } from '../../nats-wrapper';

it('return 404 if the ticket is not found', async () => {
  await request(app)
    .post('/api/orders')
    .set('Cookie', global.signin())
    .send({ ticketId: new mongoose.Types.ObjectId() })
    .expect(404);
});

it('return 400 if the ticket is allready reserved ', async () => {
  const ticket = Ticket.build({
    title: 'Test Ticket',
    price: 100,
    id: new mongoose.Types.ObjectId().toHexString(),
  });
  await ticket.save();

  const order = Order.build({
    ticket,
    userId: 'user',
    status: OrderStatus.Created,
    expiresAt: new Date(),
  });
  await order.save();

  await request(app)
    .post('/api/orders')
    .set('Cookie', global.signin())
    .send({ ticketId: ticket.id })
    .expect(400);
});

it('order done', async () => {
  const ticket = Ticket.build({
    title: 'Test Ticket',
    price: 100,
    id: new mongoose.Types.ObjectId().toHexString(),
  });
  await ticket.save();

  await request(app)
    .post('/api/orders')
    .set('Cookie', global.signin())
    .send({ ticketId: ticket.id })
    .expect(201);
});

it('publish order', async () => {
  const ticket = Ticket.build({
    title: 'Test Ticket',
    price: 100,
    id: new mongoose.Types.ObjectId().toHexString(),
  });
  await ticket.save();

  await request(app)
    .post('/api/orders')
    .set('Cookie', global.signin())
    .send({ ticketId: ticket.id })
    .expect(201);

  expect(natsWrapper.client.publish).toHaveBeenCalled();
});
